Modified Startup Script: Deploy if Maven Build is Successful

+++ start.sh
@@ -11,10 +11,17 @@ BROWSER='/c/Program\ Files\ \(x86\)/Google/Chrome/Application/chrome'
 export JPDA_ADDRESS=8000
 export JPDA_TRANSPORT=dt_socket
 
-mvn clean package
-eval find $TOMCAT_HOME/webapps -mindepth 1 -delete
-eval cp target/$ARTIFACT_NAME.war $TOMCAT_HOME/webapps/
-eval find $TOMCAT_HOME/logs  -mindepth 1 -delete
-eval find $TOMCAT_HOME/conf/Catalina/localhost -mindepth 1 -delete
-eval $TOMCAT_HOME/bin/catalina.bat jpda run &
-eval start $BROWSER http://localhost:$PORT/$ARTIFACT_NAME/$PAGE
+echo 'Wait! Building project... '
+MAVEN_LOGS=echo$(mvn clean package)
+if [[ "$MAVEN_LOGS" =~ "BUILD FAILURE" ]]
+    then
+        echo "Maven Build Failed!"
+    else
+        echo "Maven Build Success!"
+        eval find $TOMCAT_HOME/webapps -mindepth 1 -delete
+        eval cp target/$ARTIFACT_NAME.war $TOMCAT_HOME/webapps/
+        eval find $TOMCAT_HOME/logs  -mindepth 1 -delete
+        eval find $TOMCAT_HOME/conf/Catalina/localhost -mindepth 1 -delete
+        eval $TOMCAT_HOME/bin/catalina.bat jpda run &
+        eval start $BROWSER http://localhost:$PORT/$ARTIFACT_NAME/$PAGE
+fi
